Switch from Di to GetIt, create and use Notify over Bloc.
This commit solidifies the use of GetIt over
Flutter_Simple_Dependency_Injector. So far, it feels much nicer.

I also decided to ditch Bloc for now. It's one of the better state management
solutions, but I didn't really like it.

So, I created my own event system in the form for Notifier
and NotifierBuilder. All it does is allow a simple way of emitting events
and subscribing to said events, and to easily rebuild sections of the UI
without needing stateful widgets.